Refactoring: Moved client-ctrl-plane code to mapago-client.go
The client-ctrl-plane code was to much in the control-plane.go.
Control-Plane.go is much more server based than client.
